+++ docs/docs-js/api.js
@@ -1,7 +1,48 @@
 YUI.add("yuidoc-meta", function(Y) {
    Y.YUIDoc = { meta: {
-    "classes": [],
-    "modules": [],
-    "allModules": []
+    "classes": [
+        "assyncronous",
+        "browserSupport",
+        "classManipulation",
+        "form-browserSupport",
+        "formValidation",
+        "layoutFunctions",
+        "scrollingFunctions",
+        "styleManipulation"
+    ],
+    "modules": [
+        "Contact-Page",
+        "Function-JQuery",
+        "Global",
+        "Global-Mobile",
+        "Siren-Core"
+    ],
+    "allModules": [
+        {
+            "displayName": "Contact-Page",
+            "name": "Contact-Page",
+            "description": "This file contains core functions used in the Siren Framework"
+        },
+        {
+            "displayName": "Function-JQuery",
+            "name": "Function-JQuery",
+            "description": "This file contains core functions used in the Siren Framework"
+        },
+        {
+            "displayName": "Global",
+            "name": "Global",
+            "description": "This file contains the Global calls and custom functions of the project"
+        },
+        {
+            "displayName": "Global-Mobile",
+            "name": "Global-Mobile",
+            "description": "This file contains the Global calls and custom functions of the project for mobile devices"
+        },
+        {
+            "displayName": "Siren-Core",
+            "name": "Siren-Core",
+            "description": "This file contains core functions used in the Siren Framework"
+        }
+    ]
 } };
 });