+++ js/jsSrc/functions.js
@@ -3,52 +3,64 @@
 * @module Siren-Core
 */
 
+
 /**
-Methods related to manipulating classes in the HTML
-@class classManipulation
+Methods related to testing and addition browser support
+@class browserSupport
 @static
 */
 
-/**
-@method hasClass
-@param element {object} The target element
-@param className {string} The class name to check against
-@return boolean
-*/
-function hasClass(element, className) {
-    return element.className && new RegExp("(^|\\s)" + className + "(\\s|$)").test(element.className);
-}
-
 
 /**
-Used to add a class "active" to triggers and related targets
-@method activeState
-@param trigger {string} The name of the trigger
-@param target {string} The name of the target
+Test if a css selector can be used
+@method upportsSelector
+@param selector {string} 
 */
-function activeState(trigger1, target1){
-  var trigger = document.getElementById(trigger1);
-  var target = document.getElementById(target1);
-  var triggerClasses, targetClasses;
-
-  trigger.onclick = function(){
-    //Retrieve class name list for each object
-    triggerClasses = trigger.className;
-    targetClasses = target.className;
-
-    if(hasClass(trigger, 'active')){
-
-      //Remove active state from both elements
-      trigger.setAttribute("class", triggerClasses).replace('active',triggerClasses);
-      target.setAttribute("class", targetClasses).replace('active',targetClasses);
+function supportsSelector(selector) {
+    var doc = document,
+        el = doc.createElement('style'),
+        supported = false,
+        theRules;
+    
+    // IE seems to need a type to recognise a stylesheet.
+    el.type = 'text\/css';
+    
+    // This ASSUMES that IE will always give stylesheets a styleSheet method.
+    // Watch this space for errors.
+    if (el.styleSheet) {
+        el.styleSheet.cssText = selector + '{}';
+        // May as well save some typing.
+        theRules = el.styleSheet.rules;
+        // IE7 and 8 map '::before' to ':before' so we can't simply 
+        // check that our selector is the same as the returned one.
+        // The selectorText of any unrecognised selector is 'UNKNOWN' and 
+        // unrecognised Pseudo-elements come back as ':unknown', so we can
+        // check for that.
+        supported = (theRules && theRules[0].selectorText && theRules[0].selectorText.toLowerCase().indexOf('unknown') < 0);
+    } else {
+    
+        // Standards-based browsers need the stylesheet to be appended to the
+        // DOM, but they will allow us to simply give the style tag some text.
+        el.appendChild(doc.createTextNode(selector + '{}'));
+        doc.body.appendChild(el);
+        supported = !!el.sheet.cssRules.length;
+        doc.body.removeChild(el);
     }
-    else{
-      //Add active to existing class list
-      trigger.setAttribute("class", " active " + triggerClasses);
-      target.setAttribute("class", " active " + targetClasses);
-    }
-  };
+
+    // Clean up after ourselves and give us the results.    
+    el = null;
+    return supported;
 }
+
+/**
+Checks if touch device
+@method is_touch_device
+@param url {string} 
+*/
+function is_touch_device() {
+  return 'ontouchstart' in window || 'onmsgesturechange' in window;
+}
+
 
 /**
 Polyfill for querySectorAll
@@ -72,6 +84,37 @@ if (!document.querySelectorAll) {
 
 
 /**
+Methods related to manipulating classes in the HTML
+@class classManipulation
+@static
+*/
+
+
+/**
+Used to add a class "active" to triggers and related targets. ie8+
+@method activeState
+@param trigger {string} The name of the trigger
+*/
+function activeState(trigger) {
+  var targetName = trigger.getAttribute('href').replace('#', '');
+  var target = document.getElementById(targetName);
+  
+  if (!target) {
+    return;
+  }
+  
+  trigger.classList.toggle('active');
+  target.classList.toggle('active');
+}
+
+
+/**
+Methods related to manipulating style manipulation
+@class styleManipulation
+@static
+*/
+
+/**
 Used to remove an inline propert
 @method removeStyle
 @param elem {object} 
@@ -84,3 +127,33 @@ function removeStyle(elem,style){
     elem.style.removeAttribute(style);
   }
 }
+
+/**
+Methods related to assyncronous calls
+@class assyncronous
+@static
+*/
+
+/**
+Used to assyncronously load a javascript file. Based on article by Chris Lu (https://chris.lu/article/read/506de698268c420f0d000004)
+@method loadScript
+@param src {string} The path to the file from target HTML file
+@param callback {function} 
+*/
+
+function loadScript(src, callback){
+  var scriptTag = document.createElement('script'); 
+  scriptTag.src = src; 
+  scriptTag.type = 'text/javascript'; 
+  scriptTag.async = true; 
+  var headTag = document.getElementsByTagName('head')[0];
+  headTag.appendChild(scriptTag);
+
+  
+  scriptTag.onload = function(){
+    if(callback){
+      callback();
+    }
+  };
+}
+