+++ js/jsSrc/global.js
@@ -2,3 +2,4 @@
 * This file contains the Global calls and custom functions of the project
 * @module Global
 */
+